Move filemanager outside tools/utils.py to avoid an arcane import tracy deadlock
